put routes and auth middleware in separate files for better organization
